Improved tests, creating version 1.2.0
