+++ filterpy/stats/stats.py
@@ -24,6 +24,7 @@ from math import cos, sin
 from matplotlib.patches import Ellipse
 import matplotlib.pyplot as plt
 import numpy as np
+from numpy.linalg import inv
 import random
 import scipy.linalg as linalg
 import scipy.sparse as sp
@@ -40,6 +41,63 @@ try:
     multivariate_normal.logpdf(1, 1, 1, allow_singular=True)
 except:
     _support_singular = False
+
+
+
+def _validate_vector(u, dtype=None):
+    # this is taken from scipy.spatial.distance. Internal function, so
+    # redefining here.
+
+    u = np.asarray(u, dtype=dtype).squeeze()
+    # Ensure values such as u=1 and u=[1] still return 1-D arrays.
+    u = np.atleast_1d(u)
+    if u.ndim > 1:
+        raise ValueError("Input vector should be 1-D.")
+    return u
+
+
+def mahalanobis(x, mean, cov):
+    """ Computes the Mahalanobis distance between the state vector x from the
+    Gaussian  `mean` with covariance `cov`.
+
+
+    Parameters
+    ----------
+    x : (N,) array_like
+        Input state vector
+    mean : (N,) array_like
+        mean of multivariate Gaussian
+    cov : ndarray
+        covariance of the multivariate Gaussian
+
+
+    Returns
+    -------
+    mahalanobis : double
+        The Mahalanobis distance between vectors `x` and `mean`
+
+
+    Examples
+    --------
+    >>> mahalanobis(x=3., mean=3.5, cov=4.**2) # univariate case
+    0.125
+    >>> mahalanobis([1., 2], [1.1, 3.5], [[1., .1],[.1, 13]])
+    0.42533327058913922
+    """
+
+    x = _validate_vector(x)
+    mean = _validate_vector(mean)
+
+    if x.shape != mean.shape:
+        raise ValueError("length of input vectors must be the same")
+
+    y = x - mean
+    S = np.atleast_2d(cov)
+
+    # residual y is now 1D, so no need to transpose, and y @S^ @ y is a scalar
+    # so no array indexing required to get result
+    dist = np.dot(y, inv(S)).dot(y)
+    return np.sqrt(dist)
 
 
 
@@ -740,6 +798,8 @@ def NESS(xs, est_xs, ps):
     for x, p in zip(est_err, ps):
         ness.append(np.dot(x.T, linalg.inv(p)).dot(x))
     return ness
+
+
 
 
 if __name__ == '__main__':
